made a container for fetch and state, and a component for rendering
